+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.17.17] - 2026-06-25
+
+### Added
+
+- Allow audio export as WAV independent from the recording format for iNat interoperability
+
+### Changed
+
+- Sort species chips in session card alphabetically and update top species logic
+
 ## [0.17.16] - 2026-06-24
 
 ### Added